// Frontend/src/interfaces/CompanyOverview.ts
export interface CompanyOverview {
  longName: string;
  symbol: string;
  exchange: string;
  currency: string;
  sectorDisp: string;
  industryDisp: string;
  website: string;
  city: string;
  country: string;
  fiftyTwoWeekRange: string;
  marketCap: string;
  volume: string;
  open: string;
  previousClose: string;
  dayLow: string;
  dayHigh: string;
  totalRevenue: string;
  ebitda: string;
  profitMargins: string;
  grossMargins: string;
  operatingMargins: string;
  revenueGrowth: string;
  earningsGrowth: string;
  fiftyTwoWeekHigh: string;
  fiftyTwoWeekLow: string;
}